Update ImagePicker.md
The usage example for ImagePicker contains improper indentation that may be confusing to some readers. Originally, the indentations in the SwiftUI sample code in this file implied that `Image` belonged within a closure of `ImagePicker`; however, this creates an issue. `Image` should not be within a closure of `ImagePicker`, hence the change in lines 9-19.